Initial Upload
This is an extended Version of Task 4; intended to be used as a base for
task 5.
Compiler is set to "Antenna", since "Framework" didn't work and was
bypassed anyways.
(Removed code related to the Framework-project.)
Help texts are now stored in the plugin list itself, no longer in the
source code.
Other small changes.