Big Reworking: Step three (Part 1)
The entire code of each module had been moved to the aspect; not only
the code to call it.
This will make the interface uneccesary in the future.

Due to the high number of features only the first half had been done
now. The rest will follow in Step three (Part 2).